also write DataChannelOpen message on negotiated data channels
