+++ rtc/src/peer_connection/mod.rs
@@ -310,6 +310,7 @@ use shared::error::{Error, Result};
 use shared::util::math_rand_alpha;
 use std::collections::HashMap;
 use std::time::Instant;
+use crate::peer_connection::message::internal::{FlushMessage, RTCMessageInternal, TaggedRTCMessageInternal};
 
 /// Builder for creating RTCPeerConnection instances.
 ///
@@ -2234,4 +2235,35 @@ where
             .stats
             .snapshot_with_selector(now, selector)
     }
+
+    /// `flush` sends a signal that indicates when all previous messages on the given data channel
+    /// are finished sending.
+    /// After calling `flush`, a future call to `poll_flush` will emit a signal
+    /// with the same `id` indicating that all socket messages corresponding to the previous
+    /// data channel messages have been delivered via `poll_write`.
+    pub fn flush(&mut self, id: FlushId) -> std::result::Result<(), Error> {
+        let mut endpoint_handler = self.get_endpoint_handler();
+        use sansio::Protocol;
+        endpoint_handler.handle_write(TaggedRTCMessageInternal {
+            now: Instant::now(),
+            transport: Default::default(),
+            message: RTCMessageInternal::Flush(FlushMessage::new(id))
+        })
+    }
+
+    /// `poll_flush` will return the signal given to `flush` after all previous
+    /// data channel messages have been delivered via `poll_write`.
+    pub fn poll_flush(&mut self) -> Option<FlushId> {
+        self.pipeline_context.flush_outs.as_mut()
+            .and_then(|flush_outs| flush_outs.pop_front())
+    }
+}
+
+
+/// A unique identifier for a flush signal on a data channel
+#[derive(Debug, Clone)]
+pub struct FlushId {
+    /// a caller-chosen value that will be presented again when the flush signal is eventually polled
+    pub flush_id: i64,
+    pub data_channel_id: RTCDataChannelId
 }
